Create modules folder
