+++ index.html
@@ -215,7 +215,9 @@
       </div>
     </footer>
   </div>
-  <script src="script.js"></script>
+  <script src="./modules/topFunction.js"></script>
+  <script src="script.js" type="module"></script>
+  <script src="./modules/menuWindow.js"></script>
 </body>
 
 </html>
