+++ modules/menuWindow.js
@@ -1,0 +1,12 @@
+const menu = document.querySelector('.menu-dropdown');
+
+function openMenu() {
+  menu.style.visibility = 'visible';
+}
+
+function closeMenu() {
+  menu.style.visibility = 'hidden';
+}
+
+openMenu()
+closeMenu()
